more change more lore more sprites

+++ code/game/objects/structures/wrecks.dm
@@ -13,11 +13,33 @@
 /obj/structure/wreck/car/truck
 	name = "wrecked truck"
 	desc = "An old pre-war truck, rusted and destroyed with age and weathering. Its seems repairable."
-	icon = 'icons/fallout/trash.dmi'
+	icon = 'icons/fallout/objects/structures/rubish.dmi'
 	icon_state = "car_rubish13"
 	bound_width = 64
 
 /obj/structure/wreck/car/truck/attacked_by(obj/item/I, mob/living/user, params)
+	if(I.tool_behaviour == TOOL_WELDER)
+		I.play_tool_sound(src)
+		user.visible_message("<span class='notice'>[user] starts preparing the [src] for a makeover...</span>", \
+							"<span class='notice'>You start preparing the [src] for a makeover...</span>")
+		if(!I.use_tool(src, user, 50))
+			return
+		playsound(get_turf(src), 'sound/items/deconstruct.ogg', 50, 1)
+		user.visible_message("<span class='notice'>[user] dexterly opens up [src]'s space .</span>", \
+							"<span class='notice'>You dexterly open up [src]'s space.</span>")
+		new /obj/item/mecha_parts/chassis/normalvehicle/pickuptruck(loc)
+		qdel(src)
+		return
+	. = ..()
+
+/obj/structure/wreck/car/jeep
+	name = "wrecked jeep"
+	desc = "An old pre-war truck, rusted and destroyed with age and weathering. Its seems repairable."
+	icon = 'icons/fallout/objects/structures/rubish.dmi'
+	icon_state = "car_rubish14"
+	bound_width = 64
+
+/obj/structure/wreck/car/jeep/attacked_by(obj/item/I, mob/living/user, params)
 	if(I.tool_behaviour == TOOL_WELDER)
 		I.play_tool_sound(src)
 		user.visible_message("<span class='notice'>[user] starts preparing the [src] for a makeover...</span>", \
@@ -32,14 +54,14 @@
 		return
 	. = ..()
 
-/obj/structure/wreck/car/jeep
-	name = "wrecked jeep"
+/obj/structure/wreck/car/highwayman
+	name = "wrecked highwayman"
 	desc = "An old pre-war truck, rusted and destroyed with age and weathering. Its seems repairable."
-	icon = 'icons/fallout/trash.dmi'
-	icon_state = "car_rubish14"
+	icon = 'icons/fallout/objects/structures/rubish.dmi'
+	icon_state = "car_rubish15"
 	bound_width = 64
 
-/obj/structure/wreck/car/jeep/attacked_by(obj/item/I, mob/living/user, params)
+/obj/structure/wreck/car/highwayman/attacked_by(obj/item/I, mob/living/user, params)
 	if(I.tool_behaviour == TOOL_WELDER)
 		I.play_tool_sound(src)
 		user.visible_message("<span class='notice'>[user] starts preparing the [src] for a makeover...</span>", \
@@ -49,7 +71,51 @@
 		playsound(get_turf(src), 'sound/items/deconstruct.ogg', 50, 1)
 		user.visible_message("<span class='notice'>[user] dexterly opens up [src]'s space .</span>", \
 							"<span class='notice'>You dexterly open up [src]'s space.</span>")
-		new /obj/item/mecha_parts/chassis/normalvehicle/pickuptruck(loc)
+		new /obj/item/mecha_parts/chassis/normalvehicle/highwayman(loc)
+		qdel(src)
+		return
+	. = ..()
+
+/obj/structure/wreck/car/corvega
+	name = "wrecked corvega"
+	desc = "An old pre-war truck, rusted and destroyed with age and weathering. Its seems repairable."
+	icon = 'icons/fallout/objects/structures/rubish.dmi'
+	icon_state = "car_rubish16"
+	bound_width = 64
+
+/obj/structure/wreck/car/corvega/attacked_by(obj/item/I, mob/living/user, params)
+	if(I.tool_behaviour == TOOL_WELDER)
+		I.play_tool_sound(src)
+		user.visible_message("<span class='notice'>[user] starts preparing the [src] for a makeover...</span>", \
+							"<span class='notice'>You start preparing the [src] for a makeover...</span>")
+		if(!I.use_tool(src, user, 50))
+			return
+		playsound(get_turf(src), 'sound/items/deconstruct.ogg', 50, 1)
+		user.visible_message("<span class='notice'>[user] dexterly opens up [src]'s space .</span>", \
+							"<span class='notice'>You dexterly open up [src]'s space.</span>")
+		new /obj/item/mecha_parts/chassis/normalvehicle/corvega(loc)
+		qdel(src)
+		return
+	. = ..()
+
+/obj/structure/wreck/car/buggy
+	name = "wrecked buggy"
+	desc = "An old pre-war truck, rusted and destroyed with age and weathering. Its seems repairable."
+	icon = 'icons/fallout/objects/structures/rubish.dmi'
+	icon_state = "car_rubish17"
+	bound_width = 64
+
+/obj/structure/wreck/car/buggy/attacked_by(obj/item/I, mob/living/user, params)
+	if(I.tool_behaviour == TOOL_WELDER)
+		I.play_tool_sound(src)
+		user.visible_message("<span class='notice'>[user] starts preparing the [src] for a makeover...</span>", \
+							"<span class='notice'>You start preparing the [src] for a makeover...</span>")
+		if(!I.use_tool(src, user, 50))
+			return
+		playsound(get_turf(src), 'sound/items/deconstruct.ogg', 50, 1)
+		user.visible_message("<span class='notice'>[user] dexterly opens up [src]'s space .</span>", \
+							"<span class='notice'>You dexterly open up [src]'s space.</span>")
+		new /obj/item/mecha_parts/chassis/normalvehicle/buggy(loc)
 		qdel(src)
 		return
 	. = ..()
@@ -274,6 +340,30 @@
 	bound_height = 64
 	icon = 'icons/obj/vehicles/rustybus.dmi'
 	icon_state = "pulowski"
+
+/obj/structure/wreck/truck1
+	name = "wrecked truck"
+	desc = "An old pre-war vehicle, rusted and destroyed with age and weathering."
+	bound_width = 192
+	bound_height = 64
+	icon = 'icons/obj/vehicles/rustybus.dmi'
+	icon_state = "truck1"
+
+/obj/structure/wreck/truck2
+	name = "wrecked truck"
+	desc = "An old pre-war vehicle, rusted and destroyed with age and weathering."
+	bound_width = 192
+	bound_height = 64
+	icon = 'icons/obj/vehicles/rustybus.dmi'
+	icon_state = "truck2"
+
+/obj/structure/wreck/truck3
+	name = "wrecked truck"
+	desc = "An old pre-war vehicle, rusted and destroyed with age and weathering."
+	bound_width = 192
+	bound_height = 64
+	icon = 'icons/obj/vehicles/rustybus.dmi'
+	icon_state = "truck3"
 
 /obj/structure/wreck/trash/five_tires
 	name = "Tire Pile"
